+++ tools/astarte_import/lib/astarte/import.ex
@@ -26,6 +26,8 @@ defmodule Astarte.Import do
       :chars_acc,
       :object_data,
       :current_object_item,
+      :value,
+      :element_data,
       :got_device_end_fun,
       :got_end_of_value_fun,
       :got_end_of_object_fun,
@@ -247,7 +249,7 @@ defmodule Astarte.Import do
     {:ok, reception_timestamp} = fetch_attribute(attributes, 'reception_timestamp')
     {:ok, datetime, 0} = DateTime.from_iso8601(reception_timestamp)
 
-    %State{state | reception_timestamp: datetime}
+    %State{state | reception_timestamp: datetime, element_data: []}
   end
 
   defp xml_event({:startElement, _uri, _l_name, {_prefix, 'object'}, attributes}, _loc, state) do
@@ -259,7 +261,7 @@ defmodule Astarte.Import do
 
   defp xml_event({:startElement, _uri, _l_name, {_prefix, 'item'}, attributes}, _loc, state) do
     with {:ok, item_name} <- fetch_attribute(attributes, 'name') do
-      %State{state | current_object_item: item_name}
+      %State{state | current_object_item: item_name, element_data: []}
     else
       _any ->
         throw({:error, :invalid_object_item})
@@ -270,20 +272,36 @@ defmodule Astarte.Import do
     with {:ok, path} <- fetch_attribute(attributes, 'path'),
          {:ok, reception_timestamp} <- fetch_attribute(attributes, 'reception_timestamp'),
          {:ok, datetime, 0} <- DateTime.from_iso8601(reception_timestamp) do
-      %State{state | path: path, reception_timestamp: datetime}
+      %State{state | path: path, reception_timestamp: datetime, element_data: []}
     else
       _any ->
         throw({:error, :invalid_property})
     end
   end
 
+  defp xml_event({:startElement, _uri, _l_name, {_prefix, 'element'}, attributes}, _loc, state) do
+    state
+  end
+
+  defp xml_event({:endElement, _uri, _l_name, {_prefix, 'element'}}, _loc, state) do
+    %State{
+      chars_acc: chars_acc,
+      element_data: element_data
+    } = state
+
+    element_data = element_data ++ [chars_acc]
+
+    %State{state | chars_acc: nil, element_data: element_data}
+  end
+
   defp xml_event({:endElement, _uri, _l_name, {_prefix, 'property'}}, _loc, state) do
     %State{
       chars_acc: chars_acc,
+      element_data: element_data,
       got_end_of_property_fun: got_end_of_property_fun
     } = state
 
-    state = got_end_of_property_fun.(state, chars_acc)
+    state = got_end_of_property_fun.(state, chars_acc || element_data)
 
     %State{state | chars_acc: nil, reception_timestamp: nil}
   end
@@ -292,10 +310,11 @@ defmodule Astarte.Import do
     %State{
       chars_acc: chars_acc,
       object_data: object_data,
+      element_data: element_data,
       current_object_item: current_object_item
     } = state
 
-    object_data = Map.put(object_data, current_object_item, chars_acc)
+    object_data = Map.put(object_data, current_object_item, chars_acc || element_data)
 
     %State{state | chars_acc: nil, object_data: object_data, current_object_item: nil}
   end
@@ -308,18 +327,19 @@ defmodule Astarte.Import do
 
     state = got_end_of_object_fun.(state, object_data)
 
-    %State{state | reception_timestamp: nil, object_data: nil}
+    %State{state | reception_timestamp: nil, object_data: nil, element_data: nil}
   end
 
   defp xml_event({:endElement, _uri, _l_name, {_prefix, 'value'}}, _loc, state) do
     %State{
       chars_acc: chars_acc,
+      element_data: element_data,
       got_end_of_value_fun: got_end_of_value_fun
     } = state
 
-    state = got_end_of_value_fun.(state, chars_acc)
+    state = got_end_of_value_fun.(state, chars_acc || element_data)
 
-    %State{state | chars_acc: nil, reception_timestamp: nil}
+    %State{state | chars_acc: nil, reception_timestamp: nil, element_data: nil}
   end
 
   defp xml_event({:endElement, _uri, _l_name, {_prefix, 'datastream'}}, _loc, state) do